責務の分離
